requirements.txt
numpy~=1.24.3
paddleocr~=2.6.1.3
Pillow~=9.5.0
PyMuPDF~=1.20.2